Throw on FixedHash failure of String construction
I have had this bite me a few times when testing FixedHash with the
string constructor. A capital "X" in the 0x prefix, wrong number of hex
characters compared to the number of bytes the FixedHash has e.t.c. and there
was no warning. The hash had a 0 value while I was assuming it worked
fine.

Having it throw in case of error will guarantee that this won't happen again.


Former-commit-id: a7bdbe8bf997898c6f8acfca7f1f57dcab9f73ae