Reimplement InsertPointGuard to avoid LLVM ABI incompatibility.
In general, the NDEBUG flag should match cpp-ethereum and LLVM builds. But this is hard to satisfy as we usually have one system-wide build of LLVM and different builds of cpp-ethereum. This ABI incompatibility hit OSX only in release builds as LLVM is built by homebrew with assertions by default.


Former-commit-id: 64132d6d1372da14a8d987ebe5c92c0250eb3e44