Update to fix stratum failover argument handling
The stratum-failover argument handling in the existing implementation is not consistent. The existing code implements: stratum-failover and FS. However, the help screen uses both --stratum-failover and --failover-stratum. --failover-stratum is not actually implemented in the code, only the abbreviation FS is. This change modifies the code to use SF and --stratum-failover to be more consistent across the board.

Former-commit-id: 5e63e339f45eb1c581b5d242faeca4d00f086507