Swatting the invalid shares... with final optimization
Host side threads manage shared objects with mutex
locks. Fix lock use to avoid host software race condition
where stratum updates the current work package, while submit
is still reporting a valid share for the previous work
package. These two methods are called asynchronusly
by different host threads.

Since the pool message is composed from parts
of the CUDA provided solution and the host side work
package, a mismatch occurs.

There exists a more obscure race apportunity where the CUDA
code has been flagged to reset for a new job, but has just
found a solution and queued it for processing prior to
responding to the reset flag. Again, causing the submit
function to report a old solution for a new work package.
Fix this by discarding solutions produced post CUDA reset.

Both race conditions would produce an invalid share sent to
the pool.

And where this all started... we eliminate the redundant
re-evaluation of results, and send all solutions if
more than one is found by a CUDA thread block.

Oh! And optimize the job context switch. When new work
comes in the farm must reset all GPUs with the new
problem to solve. Previously it would iterate the GPUs
one at a time, resetting it, and waiting for it to
reset before moving on to the next. Improving on this
we split the operation in 2 passes. 1st we reset all
the GPUs (a non-blocking operation), then we wait
for each to reset. This should reduce the switch time
from the sum of all GPU reset times down to the time
it takes the slowest GPU to reset.