No duplicate evals, no invalid shares, minimal diff
This accomplishes the initial goal of eliminating the double
evaluation of solutions as well as supports the virtually
impossiblecase of a single thread grid finding more than one
solution. The goal this time around was not to monkey with
the threadmodel. This doesn't alter or resequence any software
locks. I'dbet that if this one locks up, then the master
code probably does as well!

As it turns out the low level C++ GPU result handler knows
when it's sending invalid solutions up the stack. It just
had no way of telling stratum about it. It seems that is
the whole reason a copy of the previous job is held and the
rechecking of the solution was done in submit. Someone
figured where these rare apparent invalids were coming from
and added a re-evaluation in submit, if that passed the share
is submitted normally, otherwise a second re-evaluation is
done againd the previous job. If that passes it is submitted
knowingly as a stale share! Expedient but inneficient at the
worst time.

The calculation of a light hash in the re-evaluation is
compute intensive and delays the sending of the share to the
pool. Not sure why you'd knowingly submit a stale share, but
there it is.

In order to keep the model as intact as possible, we simply
add a bool to the solution tuple that will hold the stale
status. It can then be conveniently passed up to stratum.
The submit method will no longer need to recheck the solution
to determine whether its for the current or last job. To keep
tings as close to the way they were, submit still sends stale
shares.

As it turns out this results in a cleaner design.