Fix scrambed console out put
Console out put formatting is protected from reentrant though I'm
not sure it needs to be? C++ 11 states that iostreams are thread
safe but they do not guarantee that the stream data won't be
intertwined.

Include the call to iostream cerr within the guard. This should
clear up the occasionally mangled console output that often
manifests as 2 lines concatenated to a single folled by a blank
line.