Get rid of failed shares... again!
What causes failed shares in the cuda miner? They
occur when the miner's work package changes between
the time it is 1st obtained at the top of its workloop,
and the time when a solution is found and the work
package is used again.

Make sure that doesn't happen by only getting new
work at the top of the work loop as is the case for
CLMiner.