Found the root cause... at last!
Turns out the x_work guard in the miner class was
not needed at all if a simple rule was followed. In
fact it was doing cuda more harm than good!

Here's what was happening, the miner thread was running
along searching for solutions and has just found one in
its report() method and needs to verify it against the
current job. To do that it calls work to create
a fresh copy of the current work. That call would
block on the x_work guard.

If a stratum notification is received precisely at
that time, before the miner gets a chance to grab the
guard, stratum notifies the farm of new work, the farm
notifies the miners, and calling the miner setWork
method grabs the guard, changes the work package,
and releases the lock. The miner thread then gets the
lock and retrieves the updated work which no longer
matches the solution its trying to submit. The
stratum thread is able to interrupt the miner threads
since it is driven by the asynchronous arrival of
a stratum packet.

The miner x_work guard is not required if miners
ONLY copy the work package at the top of the
work loop, and use reference wherever else they need
to look at current work.

The cuda miner wasn't playing by that rule and was
paying the price, not to mention the penalty it
was imposing on the CL miner, with the common and
unneccessary lock.