Lock free stratum vs. miner synchronization
I really think we need to get rid of this miner x_work
mutex. Though it successfully guards against concurrent
use of the work package while it is being updated, it
also causes a collision of the stratum and miner threads
every time a new job is posted by stratum. This might
be negligible if it didn't cause a whole lot of OS
dispatcher scheduling overhead, which is not cheap.