Fix ethereumstratum when job number is >8 chars
The current code assumes the incoming job number will always be 8 hex
characters long; this currently breaks with nicehash, which is sending
16 hex character job numbers like '000000047fa89505'.

This commit fixes the issue by storing the incoming job number length in
WorkPackage and Solution, then using it to truncate the stored job
number back to whatever the original length was.

It also fixes a closely issue where the padding, which added 56 `0`s,
was padding out to 72 characters; the fix here resizes directly to 64
characters, whatever the input size.