No need to call the randomizer with each new job.
We don't want to delay a job switch with an expensive
random number generation. Besides this approach does
not guarantee that two GPU will not be given overlapping
search ranges (highly unlikely but possible).

Instead we give each GPU its own unique 2^40 nonce
range to search and we do it very efficiently.


Former-commit-id: 9e042007c5d71c9f3f6a4603077bc4329bff00d4