CUDA: Eliminate software evals completely!
Given that we trust our GPUs there is no good reason
to recalculate the mix hash in software. The GPU has
already calculated it in order to find the nonce.

So, we get CUDA to provide us with the mix hashes
corresponding to each found solution, and send that
off to the pool insted of the recalculated one which
would be identical anyway!

What was it? Another 6ms. saved, I think.