Exit on cuda run-time error
These errors typically occur on overclocked
CUDA GPUs and are presently unrecovered.
The often result in a miner that appears
to be mining but is finding no shares.

These often can be recovered with a miner
restart. Give the opportunity for the user
to do so via script by exiting on such errors.