Get rid of unused or inconsistent protocol method
The m_running variable and associated isRunning method
is only set to false in stratum where it was actually
used incorrectly!

In my view, if a pool protocol client is instantiated then
by definition it is running. The m_connected and m_registered
variables are sufficient to manage the protocol state.

The termination of the doStratum loop in was conditioned
by the client's isRunning state, which was only ever
innappropriately set to false on a disconnect.

The loop now will never terminate, which is a problem
for controlled shutdown. I believe the coming conversion
of stratum to a full fledged pool protocol client
will correct the problem.