Implement URI support for specifying pool all pool attributes.
Many of the current parameter options that are no longer necessary
are flagged at deprecated, but remain functional for backwards
compatibility. Presumably support for these would be dropped in
a future release.

In short, the new -P parameter allows specifying up to 6 pool
endpoints of the form scheme://[[user[:password]@]host:port