Fixed benchmark inner mean calculation when several speeds are equal
Previously, the speed results were stored in a map. This had the advantage of being sorted automatically, which made access to min and max speeds very easy
However, it had the logical problem of map collision in cases where several runs had identical speeds. While this would not affect min and max speeds, it did affect inner mean calculations (duplicate speeds count only once)

Additionally, the map data structure was not used sensibly; The rate was used as key, but also be the only attribute ever accessed from the value. Thus, the data can be represented cleaner and more accurately with the usage of a list
This commit implements the required changes to fix the inner mean calculation.