improving output/behavior of CLMiner::configureGPU() if a subset of the available GPUs is selected
What I think CLMiner::configureGPU() intends to do: 
Check for each OpenCL device [GPU] that is has sufficient Memory in order to contain the DAG file.

What it currently does:
It checks if the first (and only the first) OpenCL device in the given platform has sufficient memory.
It ignores any GPU selection settings. If the first device does not fulfill the memory requirement, 

What it does after including my change:
It checks every OpenCL device in the given platform, and only quits if none of them fulfill the memory requirement.