Use FIFO queue to track submission timings
More precise tracking of submissions times using a fifo queue. Till now
the response time tracking was always referred to the "last" submission
which could be misleading when two or more submissions are posted in
sequence before the first one receives a response.
This behavior has more evidence on pools with low/adjustable diff