Stratum mode autodetection
Eliminates the hassle of knowing in advance which is the stratum mode implemented by the pool. No more stratum/1/2+tcp:// but only stratum+tcp://

Added support for invalid connections. Whenever a connection cannot be established due to login errors (invalid worker or bad address) it's marked as Unrecoverable and won't be used again.