Auto-Stratum refinement
As per user report here
https://github.com/ethereum-mining/ethminer/pull/1333#issuecomment-402641473

There is an issue with f2pool.com which, in case of wrong login technique, instead of answering back with a Json error message DOES NOT RESPOND AT ALL and begins, after a while, to send jobs.

This PR amends the behavior introducing a timeout on first responses.

It also addresses #1339 which may be caused by the sinchronous cancel of the timer which is triggering the disconnection.