Generate DAG only at new epoch.
As the DAG only depends on the epoch we don't have to destroy all mining
threads and delete the DAG if we change a connection.
Keep the mining threads alive but suspend them while connection change.
If the new connection has a different epoch DAG will be generated when
mining is resumed.

This should result in faster pool failover switches or faster connection
switches via API as the DAG will be only generated if we get a different
epoch.