A method for nearly instantaneous hash rate measurements.
Instead of calculating hash rate at the collector's polling pass,
calculate the current hash rate of the last n batches in real
time. N is chosen to provide relatively long (.5 secs) GPU search
loop driven calculation interval. For AMD n=1 works well, and for
Nvidia which uses short batches N=128.

This PR does not affect the collector thread or the display loop,
but the collector thread may not be needed.