Provide more granular control of logging options
The verbosity level was intended to indicate severity of
a log entry. Over time we've used it inappropriately to
enable various log options. This has led us to a situation
where it is impossible to enable some options without
also enabling others.

This changes the global verbosity setting to be a set of
bits, each indicating if a log feature is enabled.

The following choices are defined and spelled out in the
help.

 #define LOG_SWITCH_TIME 1
 #define LOG_JSON 2
 #define LOG_PER_GPU 4
 #define LOG_DEBUG 8