CUDA search loop optimization - pass 2!
It makes more sense to do non critical things like calculating
hash rates and submitting solutions while the CPU is idle instead
of when the GPU is idle. But the added complexity could not
be justified based on analysis.

That analysis was based on submit time measurements of less than
1 microsecond, which is indeed the case when using the --noeval
option. However software eval is on by default. Recalculating
the mix hash in software takes 4-5 milliseconds on a reasonably
fast CPU. This imposes a much more substantial delay restarting
a stream, and has a cascading effect delaying all other streams.

This approach is not much different from previous implementations.
It merely saves all found solutions till all cuda streams searches
have been restarted and are not waiting for the CPU. Once that
is done, the CPU is free to handle hash rate and submits without
delaying cuda streams.