Faster exit with CUDA devices.
A call to cudaDeviceReset() needs in my environment between 1 and 5 seconds.
When stopping ethminer, Farm::stop() clears the m_miners list.
This involves that **sequential** the destructor of each m_miner element is called.
As the destructor of CUDAMiner waits till mining of the thread has stopped that
way can cause exiting needs up to 5 seconds per card.

With this patch the mining threads gets async informed to stop
which causes a faster exit time.