Remove accounting for stale shares
Measurement of stales at miner's side has no meaning and in 99% of cases does not match with stales accounted by pool.
In addition account stales per GPU is a nonsese: staleness is a matter of jobs, their frequency, and network latency. If a GPU produces more stales than another it's only a matter of bad luck: not a statistical data to rely on.