Allow passing "." as username or workername using %2e.
As we split username and workername by "." we have to do it before
an urlDecode() is called as otherwise it is not clear which part is
username and which is workername.

Previous this commit:
-P username%2e246891.rigname%2e01:x@HOST:PORT
		Username = username, Workername = 246891.rigname.01
-P username%2erigname:x@HOST:PORT
		Username = username, Workername = rigname

With this commit
-P username%2e246891.rigname%2e01:x@HOST:PORT
		Username = username.246891, Workername = rigname.01
-P username%2erigname:x@HOST:PORT
		Username = username.rigname, Workername =

But: this commit breaks a syntax of
	username%2erigname:x@HOST:PORT

as now this results in a
	Username = username.rigname
against previous
	Username = username, Workername = rigname

Thinking nobody used such a syntax.

This commit should fix
	https://github.com/ethereum-mining/ethminer/issues/1687
	https://github.com/ethereum-mining/ethminer/issues/1075