update hunter link for issue #2290
This PR is used to fix the issue error: downloading 'https://dl.bintray.com/boostorg/release/1.66.0/source/boost_1_66_0.7z' failed #2290
The original hunter is ruslo/hunter is not maintained. The link of boost included in it is unavailable now.

So, update the cpp_pm/hunter to fix the boost link issue.